# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CommerceFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserList {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ProfileUpdate {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  customer: 'customer',
  admin: 'admin',
} as const;

export interface UserUpdate {
  name?: string;
  role?: UserUpdateRole;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  stock: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  imageUrl: string;
  images?: string[];
  rating: number;
  reviewCount: number;
  isFeatured?: boolean;
  createdAt: string;
}

export interface ProductList {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  stock: number;
  categoryId: number;
  imageUrl: string;
  images?: string[];
  isFeatured?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  /** @nullable */
  comparePrice?: number | null;
  stock?: number;
  categoryId?: number;
  imageUrl?: string;
  images?: string[];
  isFeatured?: boolean;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount: number;
}

export interface CategoryInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CategoryUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CartItem {
  productId: number;
  name: string;
  price: number;
  imageUrl: string;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  productId: number;
  name: string;
  price: number;
  imageUrl: string;
  quantity: number;
  subtotal: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  packed: 'packed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  items: OrderItem[];
  total: number;
  shippingAddress: string;
  createdAt: string;
}

export interface OrderList {
  orders: Order[];
  total: number;
  page: number;
  totalPages: number;
}

export interface OrderInput {
  shippingAddress: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  packed: 'packed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface Review {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  productId: number;
  rating: number;
  comment: string;
  createdAt: string;
}

export interface ReviewInput {
  rating: number;
  comment: string;
}

export interface WishlistItem {
  id: number;
  userId: number;
  productId: number;
  product: Product;
  addedAt: string;
}

export interface DashboardStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalCustomers: number;
  revenueGrowth: number;
  orderGrowth: number;
  lowStockCount: number;
  pendingOrders: number;
}

export interface RevenueDataPoint {
  month: string;
  revenue: number;
  orders: number;
}

export interface TopProduct {
  productId: number;
  name: string;
  imageUrl: string;
  totalSold: number;
  revenue: number;
}

export interface CategoryDistribution {
  categoryName: string;
  count: number;
  percentage: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  order_placed: 'order_placed',
  order_shipped: 'order_shipped',
  order_delivered: 'order_delivered',
  user_registered: 'user_registered',
  review_posted: 'review_posted',
} as const;

export type ActivityItemMetadata = { [key: string]: unknown };

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  createdAt: string;
  metadata?: ActivityItemMetadata;
}

export interface InventoryAlert {
  productId: number;
  name: string;
  stock: number;
  imageUrl: string;
  threshold?: number;
}

export type ListProductsParams = {
page?: number;
limit?: number;
search?: string;
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
minPrice?: number | null;
/**
 * @nullable
 */
maxPrice?: number | null;
sort?: ListProductsSort;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  newest: 'newest',
  oldest: 'oldest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating: 'rating',
} as const;

export type ListOrdersParams = {
page?: number;
/**
 * @nullable
 */
status?: string | null;
};

export type ListUsersParams = {
page?: number;
/**
 * @nullable
 */
search?: string | null;
};