# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CommerceFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  items: OrderItem[];
  total: number;
  shippingAddress: string;
  createdAt: string;
}